requirements.txt
audioread==2.1.6
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0.post1812010730
idna==2.8
iso-639==0.4.5
iso3166==0.9
isodate==0.6.0
progressbar2==3.39.2
pyacoustid==1.1.5
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.7.3
PySocks==1.6.8
python-utils==2.3.0
requests==2.22.0
rsa==4.0
six==1.12.0
streamlink==1.0.0
uritemplate==3.0.0
urllib3==1.25.3
websocket-client==0.54.0
ruamel.yaml==0.15.87
youtube-dl==2019.11.22
